Source code release for version 0.1.104-beta
on-behalf-of: @alternetsoft <yevgeni.zolotko@alternetsoft.com>

+++ Alternet.UI/Controls/Grid/XamlGridLengthSerializer.cs
@@ -1,12 +1,8 @@
 #nullable disable
 
 using System;
-using System.ComponentModel;
-using System.ComponentModel.Design.Serialization;
 using System.Diagnostics;
 using System.Globalization;
-using System.IO;
-using System.Reflection;
 
 namespace Alternet.UI
 {
@@ -40,7 +36,7 @@ namespace Alternet.UI
         static internal void FromString(
                 string s,
                 CultureInfo cultureInfo,
-            out float value,
+            out double value,
             out GridUnitType unit)
         {
             string goodString = s.Trim().ToLowerInvariant();
@@ -51,7 +47,7 @@ namespace Alternet.UI
             int i;
             int strLen = goodString.Length;
             int strLenUnit = 0;
-            float unitFactor = 1.0f;
+            double unitFactor = 1.0f;
 
             //  this is where we would handle trailing whitespace on the input string.
             //  peel [unit] off the end of the string
@@ -111,7 +107,7 @@ namespace Alternet.UI
 
 //                ReadOnlySpan<char> valueString = goodString.AsSpan(0, strLen - strLenUnit);
                 var valueString = goodString.Substring(0, strLen - strLenUnit);
-                value = float.Parse(valueString, provider: cultureInfo) * unitFactor;
+                value = double.Parse(valueString, provider: cultureInfo) * unitFactor;
             }
         }
 
@@ -125,7 +121,7 @@ namespace Alternet.UI
 
         //  this array contains strings for unit types that are not present in the GridUnitType enum
         static private string[] PixelUnitStrings = { "in", "cm", "pt" };
-        static private float[] PixelUnitFactors =
+        static private double[] PixelUnitFactors =
         {
             96.0f,             // Pixels per Inch
             96.0f / 2.54f,      // Pixels per Centimeter
